# rust: make CRC checking optional