Summary:
Instead of computing checksums for every record, we can now compute
checksums only for records that fail proto parsing. This means that bit
flips in payloads like string fields may not be caught. But it comes
with a significant performance improvement:

  - 2.42× (±0.24×) faster for `mnist` (125 ms → 50 ms)
  - 1.20× (±0.03×) faster for `edge_cgan` (10.4 s → 8.7 s)
  - 1.23× (±0.01×) faster for `nndiv` (182 s → 148 s)

Both the main RustBoard binary and the `bench` binary now take
`--checksum`/`--no-checksum` flags, defaulting to `--no-checksum`.

Test Plan:
Unit tests extended. Benchmarked with:

```
hyperfine -w3 \
    './target/release/bench --logdir ~/testdata/edge_cgan/ {cksum}' \
    -L cksum --checksum,--no-checksum
```

wchargin-branch: rust-optional-crc